requirements.txt
ipykernel
python-dotenv
google-adk
litellm